Moved all food style sheets into a foodStyle folder
