+++ frontend/src/styles/foodStyles/foodCardStyle.module.scss
@@ -1,4 +1,4 @@
-@import './default.module.scss';
+@import '../default.module.scss';
 
 .foodCardWrapper {
   $transition-time: 0.15s;
