--- frontend/src/styles/foodsecStyles.module.scss
@@ -1,33 +1,0 @@
-@import './default.module.scss';
-
-.FoodDisplayWrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  margin-bottom: 60px;
-}
-
-.FoodMenuContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-}
-
-.titleWrapper {
-  width: 100%;
-  justify-content: center;
-  text-align: center;
-}
-
-.title {
-  color: $highlight-colour;
-}
-
-.cardDisplayWrapper {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  width: fit-content;
-}
